基本外壳完成

+++ src/renderer/views/setting/setting.js
@@ -9,21 +9,19 @@ const APP = {
   mounted () {},
   methods: {
     handleExit () {
-      this.modelVisible = true
-    },
-    handleConfirm () {
-      localStorage.clear()
-      this.$router.push({
-        path: '/signin'
+      // this.modelVisible = true
+      this.$Modal.confirm({
+        title: '安全退出',
+        content: '是否退出 Panda Tools？',
+        type: 'info',
+        onOk: async () => {
+          localStorage.clear()
+          this.$router.push({
+            path: '/signin'
+          })
+          this.$message.success('您已安全退出 Panda Tools')
+        }
       })
-      this.$message.success('您已安全退出 Panda tools')
-    },
-    handleCancel () {
-      document.getElementById('exitModal').classList.remove('fadeIn')
-      document.getElementById('exitModal').classList.add('fadeOut')
-      setTimeout(() => {
-        this.modelVisible = false
-      }, 1200)
     }
   }
 }
